// core/src/types.d.ts
export type NodeType = "text" | "variable" | "directive" | "javascript";
export interface SourceLocation {
    line: number;
    column: number;
}
export interface Node {
    type: NodeType;
    content?: string;
    name?: string;
    args?: any[];
    start?: number;
    end?: number;
    loc?: {
        start: SourceLocation;
        end: SourceLocation;
    };
    children?: Node[];
    related?: Node[];
    raw?: boolean;
}